refactor: 기존 안전 결제에서 상태로 관리하던 방식에서 어드민 계좌를 만들어서 3자 거래로 변경 및 금액 관련 부분 Double로 변경

+++ src/main/java/com/mfc/payment/domain/Payment.java
@@ -28,7 +28,7 @@ public class Payment extends BaseTimeEntity{
 	private Long paymentId;
 
 	@Column(name = "amount", nullable = false)
-	private Integer amount;
+	private Double amount;
 
 	@Column(name = "payment_status", nullable = false)
 	private String paymentStatus;
@@ -37,7 +37,7 @@ public class Payment extends BaseTimeEntity{
 	private LocalDateTime paymentDate;
 
 	@Builder
-	public Payment(String uuid, Long paymentId, Integer amount, String paymentStatus, LocalDateTime paymentDate) {
+	public Payment(String uuid, Long paymentId, Double amount, String paymentStatus, LocalDateTime paymentDate) {
 		this.uuid = uuid;
 		this.paymentId = paymentId;
 		this.amount = amount;
